inputs: do not bind keys in base inputs
Often a game includes multiple inputs of the same type (e.g. multiple
actuators or buttons). With the current system, we bind those to the
same keybinds, resulting in a keybind which does multiple things. This
can be annoying, or in the worst case even dangerous if certain motors
or actuators should not move at the same time.

To avoid this, remove the defaults in the base classes. Template games
and examples can show how to set the default keybinds in the python code
in a way that avoids duplicating keybinds (unless it's done on purpose)